pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edith-command-center"
version = "0.1.0"
description = "A local desktop assistant with voice control, automation, and multi-model Ollama support."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Abhay" }
]
dependencies = [
  "requests==2.32.5",
  "SpeechRecognition==3.15.1",
  "PyAudio==0.2.14",
  "pyautogui==0.9.54",
  "pyperclip==1.9.0",
  "pywhatkit==5.4",
  "Wikipedia-API==0.8.1",
  "phonenumbers==9.0.25",
  "pywin32==311",
  "pycaw==20251023",
  "vosk==0.3.45",
  "Pillow==11.3.0",
  "python-docx==1.2.0",
]

[project.scripts]
edith = "edith_app.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["edith_app*"]

